Removed stupid bug in parallelization of global_operations(). Double precision is now default.

+++ ANNarchy/core/Global.py
@@ -57,7 +57,7 @@ config = dict(
     'verbose': False,
     'show_time': False,
     'suppress_warnings': False,
-    'float_prec': 'single',
+    'float_prec': 'double',
     'num_threads': None, #default by os
     'paradigm': "openmp"
    }
